+++ previous-versions/sync-for-expenses-version-1/src/sdk/models/operations/updateexpensedataset.ts
@@ -11,6 +11,9 @@ export class UpdateExpenseDatasetRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "request, media_type=application/json" })
     updateExpenseRequest?: shared.UpdateExpenseRequest;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
@@ -25,12 +28,18 @@ export class UpdateExpenseDatasetRequest extends SpeakeasyBase {
  * Accepted
  */
 export class UpdateExpenseDataset202ApplicationJSON extends SpeakeasyBase {
+    /**
+     * Unique identifier for the updated sync.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "syncId" })
     syncId?: string;
 }
 
 export class UpdateExpenseDatasetResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -40,9 +49,15 @@ export class UpdateExpenseDatasetResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;