+++ previous-versions/sync-for-expenses-version-1/src/sdk/models/shared/syncfailedwebhookdata.ts
@@ -3,9 +3,16 @@
  */
 
 import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
-import { Expose, Type } from "class-transformer";
+import { Expose } from "class-transformer";
 
-export class SyncStartedWebhookData extends SpeakeasyBase {
+export class SyncFailedWebhookData extends SpeakeasyBase {
+    /**
+     * The stage of the job the sync failed.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "FailureStage" })
+    failureStage?: string;
+
     /**
      * In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
      *
@@ -73,63 +80,4 @@ export class SyncStartedWebhookData extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "syncType" })
     syncType?: string;
-}
-
-/**
- * Webhook request body used to notify that a sync has started.
- */
-export class SyncStartedWebhook extends SpeakeasyBase {
-    /**
-     * Unique identifier of the webhook event.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "AlertId" })
-    alertId?: string;
-
-    /**
-     * Unique identifier for your client in Codat.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "ClientId" })
-    clientId?: string;
-
-    /**
-     * Name of your client in Codat.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "ClientName" })
-    clientName?: string;
-
-    /**
-     * Unique identifier for your SMB in Codat.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "CompanyId" })
-    companyId?: string;
-
-    @SpeakeasyMetadata()
-    @Expose({ name: "Data" })
-    @Type(() => SyncStartedWebhookData)
-    data?: SyncStartedWebhookData;
-
-    /**
-     * A human readable message about the webhook.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "Message" })
-    message?: string;
-
-    /**
-     * Unique identifier for the rule.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "RuleId" })
-    ruleId?: string;
-
-    /**
-     * The type of rule.
-     */
-    @SpeakeasyMetadata()
-    @Expose({ name: "RuleType" })
-    ruleType?: string;
 }